Refactor page callback registration
