+++ pages/__init__.py
@@ -1,5 +1,3 @@
 """Page callback registry."""
 
-from . import upload, device_learning, data_enhancer
-
-__all__ = ["upload", "device_learning", "data_enhancer"]
+__all__ = ["greetings", "upload", "device_learning", "data_enhancer"]
