+++ pages/data_enhancer/__init__.py
@@ -1,0 +1,5 @@
+"""Data enhancer page package."""
+
+from .callbacks import register_callbacks
+
+__all__ = ["register_callbacks"]
