+++ pages/device_learning/__init__.py
@@ -1,0 +1,5 @@
+"""Device learning page package."""
+
+from .callbacks import register_callbacks
+
+__all__ = ["register_callbacks"]
